[TICKET 012] Fix CI workflow, data.ts provenance, and DashboardPage
- Switch CI from npm to Bun (oven-sh/setup-bun@v2); removes reference to
  non-existent package-lock.json that was causing the job to fail immediately
- Add clarifying header to data.ts marking it as temporary mock scaffolding,
  to distinguish it from the design reference file it was adapted from
- Extract module-level season computations (completedRaces, winnerCounts,
  seasonAvgTop3/Exact/MPE, bestRace) into src/utils/season.ts; adds empty-
  array guards to prevent NaN/undefined at the start of a season
- Fix hardcoded "100% podium" sub-label on Best Race stat card — now reads
  the actual top3Accuracy value from RACE_ACCURACY
- DashboardPage.tsx trimmed from 151 → 117 lines
